+++ lib/features/onboarding/onboarding_ui.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:animated_text_kit/animated_text_kit.dart';
 import 'package:health_track/main.dart';
+import '../homescreen/homescreen.dart';
 import '../journaling/journaling_ui.dart';
 import 'onboarding_controller.dart';
 //import '../journaling/journaling_screen.dart';
